telegram_bot/requirements.txt
aiogram==2.25.1
asyncio==3.4.3
openai
nats-py==2.5.0
python-dotenv==1.0.1